Reorganize triton directory
